+++ libraries/extractor/src/main/java/androidx/media3/extractor/ts/H263Reader.java
@@ -221,13 +221,11 @@ public final class H263Reader implements ElementaryStreamReader {
   }
 
   @Override
-  public void packetFinished(boolean isEndOfInput) {
+  public void endOfInputReached() {
     // Assert that createTracks has been called.
     checkNotNull(sampleReader);
-    if (isEndOfInput) {
-      sampleReader.onDataEnd(totalBytesWritten, /* bytesWrittenPastPosition= */ 0, hasOutputFormat);
-      sampleReader.reset();
-    }
+    sampleReader.onDataEnd(totalBytesWritten, /* bytesWrittenPastPosition= */ 0, hasOutputFormat);
+    sampleReader.reset();
   }
 
   /**