+++ libraries/extractor/src/main/java/androidx/media3/extractor/ts/H264Reader.java
@@ -183,15 +183,13 @@ public final class H264Reader implements ElementaryStreamReader {
   }
 
   @Override
-  public void packetFinished(boolean isEndOfInput) {
+  public void endOfInputReached() {
     assertTracksCreated();
-    if (isEndOfInput) {
-      seiReader.flush();
-      // Simulate end of current NAL unit and start an AUD one to trigger output of current sample
-      endNalUnit(totalBytesWritten, 0, 0, pesTimeUs);
-      startNalUnit(totalBytesWritten, NalUnitUtil.H264_NAL_UNIT_TYPE_AUD, pesTimeUs);
-      endNalUnit(totalBytesWritten, 0, 0, pesTimeUs);
-    }
+    seiReader.flush();
+    // Simulate end of current NAL unit and start an AUD one to trigger output of current sample
+    endNalUnit(totalBytesWritten, 0, 0, pesTimeUs);
+    startNalUnit(totalBytesWritten, NalUnitUtil.H264_NAL_UNIT_TYPE_AUD, pesTimeUs);
+    endNalUnit(totalBytesWritten, 0, 0, pesTimeUs);
   }
 
   @RequiresNonNull("sampleReader")