+++ libraries/extractor/src/test/java/androidx/media3/extractor/ts/TsExtractorTest.java
@@ -559,7 +559,7 @@ public final class TsExtractorTest {
     public void consume(ParsableByteArray data) {}
 
     @Override
-    public void packetFinished(boolean isEndOfInput) {
+    public void packetFinished() {
       packetsRead++;
     }